+++ service/controller/resource/prometheus/resource.go
@@ -168,6 +168,12 @@ func toPrometheus(v interface{}, config Config) (metav1.Object, error) {
 					// memory: 5Gi
 					corev1.ResourceMemory: *resource.NewQuantity(5000*1024*1024, resource.BinarySI),
 				},
+				Limits: corev1.ResourceList{
+					// cpu: 100m
+					corev1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI),
+					// memory: 5Gi
+					corev1.ResourceMemory: *resource.NewQuantity(5000*1024*1024, resource.BinarySI),
+				},
 			},
 			Retention:      config.RetentionDuration,
 			RetentionSize:  config.RetentionSize,